# Add error detection for arguments of DSU methods