## Why

ライブラリとして利用する場合は、エラーを教えてあげた方が使い勝手がいいと考えるため。
また、特別、遅くもならなかったので、PRとしてだしました。
そもそも、本家のライブラリでは、ちゃんとエラー判定がありました。

## Note: Benchmark

Current:
https://atcoder.jp/contests/atc001/submissions/22706886 307ms
https://atcoder.jp/contests/atc001/submissions/22706938 311ms

New:
https://atcoder.jp/contests/atc001/submissions/22706927 324ms
https://atcoder.jp/contests/atc001/submissions/22706922 318ms